Finally moved Effect into its own file. Changed usage of dynamicID in PowerManager to use uint. PowerManager.UsePower() will now look in the user's world for Monster instances that match the initial target id. Added speed parameter to MoveActorNormal(), which allows knockback to move actors at a reasonable speed instead of instantly.

+++ src/Mooege/Core/GS/Player/Player.cs
@@ -382,7 +382,7 @@ namespace Mooege.Core.GS.Player
         // Message handlers
         private void OnObjectTargeted(GameClient client, TargetMessage message)
         {
-            this.World.Game.PowerManager.UsePower(this, message.PowerSNO, (int)message.TargetID, message.Field2.Position, message);
+            this.World.Game.PowerManager.UsePower(this, message.PowerSNO, message.TargetID, message.Field2.Position, message);
 
             //Actor actor = this.World.GetActor(message.TargetID);
             //if (actor != null)
